* commit '70cdf1531159edb0ee64a969bf8720945ff6dec3': (25 commits)
  Reverted from sunindextype ~> long int that I missed earlier.
  Numerous small fixes to example codes and Fortran BBDOPT function interfaces.  For a longer explanation of these updates, see my comment on the 'indextype' pull request.
  SP: Updated comments in sundials_types.h
  SP: Fixing cut-paste error in sundials_types.h
  SP: Removing accidentally added IDE config file.
  SP: Change name indextype ~> sunindextype
  SP: Reverting all indextype changes in CPODES.
  SP: Reverted changes indextype ~> long int in examples where applicable.
  SP: Extended comments in sundials_types.h
  Fixed casts in <solver>_SETIIN for functions that have input types other than 'int'.
  Fixed reversion of counters from 'indextype' to 'long int', as well as a few counters that were still 'int' instead of 'long int'.  Also improved spacing somewhat.
  Fixed types for error flags and last flags in dense/band linear solver interfaces
  Fixed reversion of nge in BBD routines (both C and Fortran) to 'long int'
  Fixed conversion of 'long int' to 'indextype'
  Fixed PVEC_INTEGER_MPI_TYPE to work with indextype
  fixed int/indextype issues in nvector_parallel_grid
  SP: Reverted indextype ~> long int in Arkode, where applicable.
  SP: Revert indextype changes in CVODES where needed.
  SP: Revert indextype ~> long int where appropriate in CVODE
  SP: Reverted indextype ~> long int where appropriate in Kinsol.
  ...

+++ examples/arkode/C_serial/ark_brusselator1D_FEM_slu.c
@@ -433,7 +433,7 @@ static int f_diff(realtype t, N_Vector y, N_Vector ydot, void *user_data) {
   realtype dw = udata->dw;
 
   /* local variables */
-  long int i;
+  int i;
   realtype ul, ur, vl, vr, wl, wr;
   realtype xl, xr, f1;
   booleantype left, right;
@@ -514,7 +514,7 @@ static int f_rx(realtype t, N_Vector y, N_Vector ydot, void *user_data) {
   realtype ep = udata->ep;
 
   /* local variables */
-  long int i;
+  int i;
   realtype ul, ur, vl, vr, wl, wr;
   realtype u, v, w, xl, xr, f1, f2, f3;
   booleantype left, right;
